Finish fixing fluid requirement

+++ src/main/java/fr/frinn/custommachinery/common/component/FluidMachineComponent.java
@@ -224,7 +224,7 @@ public class FluidMachineComponent extends AbstractMachineComponent implements I
 
     /** Recipe Stuff **/
 
-    public long getRecipeRemainingSpace() {
+    public int getRecipeRemainingSpace() {
         if(!this.fluidStack.isEmpty())
             return this.capacity - this.fluidStack.getAmount();
         return this.capacity;
